Use DWORD_PTR for string returns and add Steam stub
Convert string-returning APIs from DWORD to DWORD_PTR and add proper casts/static_casts to preserve pointer-width on x64 (fixes truncation issues). Update resource/mangled signatures and related command strings to match 64-bit calling/decorated names. Define DLLEXPORT for LightMapper export. Add a minimal SteamMultiplayer stub (steam_stub.cpp, steam_stub.def) and a CMake target (CMakeLists.txt) to provide a no-op SteamMultiplayer.dll at runtime when the Steamworks SDK is unavailable. These changes improve x64 compatibility and allow the engine to run without the Steam SDK.

+++ Dark Basic Public Shared/Dark Basic Pro SDK/DarkSDKMore/Enhancements/Core.cpp
@@ -117,7 +117,7 @@ void Destructor ( void )
 char* ENHANCEMENTSSetupString ( const char* szInput )
 {
 	char* pReturn = NULL;
-	DWORD dwSize  = strlen ( szInput );
+	DWORD dwSize  = static_cast<DWORD>( strlen ( szInput ) );
 	g_pGlob->CreateDeleteString((DWORD_PTR*)&pReturn, dwSize + 1 );
 	if ( !pReturn )
 	{
